moved file loader helpers into tools
